fix(core): correct partial-run divergence detection and clippy lints
detect_divergence treated a shorter observed sequence than the journal
as nondeterminism, which contradicts the partial journal resume
guarantee (a paused or in-progress run has not yet produced every
expected activity, and that is valid, not divergence). Only a value
mismatch at a shared position, or extra activity beyond the journal
end, is a real divergence. det_divergence_code_change.rs asserted the
old, incorrect behavior; corrected it to match.

Also fixes needless_range_loop in eval.rs, needless_return in
executor.rs, collapsible_match in replay.rs, unnecessary_cast in
retry.rs, and nonminimal_bool in security_mcp_auth_tests.rs, all of
which block the workspace-wide clippy lint job.

+++ crates/ancora-core/src/retry.rs
@@ -81,7 +81,7 @@ where
                     };
                 }
                 if attempt < policy.max_attempts {
-                    let exp = (attempt - 1) as u32;
+                    let exp = attempt - 1;
                     let shift = exp.min(63);
                     let base = policy.initial_backoff_ms.saturating_mul(1u64 << shift);
                     let capped = base.min(policy.max_backoff_ms);
